Added the Resident SoundFont Synth
The Resident SoundFont Synth is initially a
renamed copy of Sf2Synth1, located on the same
web site, but in a directory outside the
WebMIDISynthHost directory.
The Resident SoundFont Synth is going to be
developed for use in other applications (such as
my Assistant Performer). The Sf2Synth1 is now
archived at its present state, pending discussions
about the Web MIDI API.